Former-commit-id: f0bbe9e89c14f54df34c42a0bf71be529cd38259

+++ Sources/Internal/Render/RenderState.cpp
@@ -29,7 +29,6 @@
 
 #include "Render/RenderState.h"
 #include "Render/RenderManager.h"
-#include "Debug/Backtrace.h"
 #include "Render/Shader.h"
 #include "Platform/Thread.h"
 #include "Utils/Utils.h"
@@ -248,7 +247,7 @@ void RenderState::Flush(RenderState * hardwareState) const
             
             uint32 minIndex = Min(currentTextureData.minmaxTextureIndex & 0x000000FF, hardwareTextureData.minmaxTextureIndex & 0x000000FF);
             uint32 maxIndex = Max((currentTextureData.minmaxTextureIndex & 0x0000FF00), (hardwareTextureData.minmaxTextureIndex & 0x0000FF00)) >> 8;
-            for(size_t i = minIndex; i <= maxIndex; ++i)
+            for(uint32 i = minIndex; i <= maxIndex; ++i)
             {
                 if(currentTextureData.textures[i] != hardwareTextureData.textures[i])
                 {
